requirements.txt
fastapi
pydantic
pydantic[email]
pyjwt
pytest
python-dotenv
requests
uvicorn
